move addToBoard from BoardService to UserService, remove redundant UserBoardService
