+++ src/main/java/com/codurance/retropolis/repositories/BoardRepository.java
@@ -6,7 +6,5 @@ public interface BoardRepository {
 
   Board getBoard(Long id);
 
-  void addToBoard(Long userId, Long boardId);
-
   Board insert(Board board);
 }
